# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyApp",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects Endpoint",
        "description": "项目列表.",
        "operationId": "list_projects_endpoint_api_v1_projects_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project Endpoint",
        "description": "创建项目.",
        "operationId": "create_project_endpoint_api_v1_projects_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/data-sources": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Data Sources Endpoint",
        "description": "列出可选数据源.",
        "operationId": "list_data_sources_endpoint_api_v1_projects_data_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DataSourceInfo"
                  },
                  "type": "array",
                  "title": "Response List Data Sources Endpoint Api V1 Projects Data Sources Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/projects/data-sources/upload-folder": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Upload Data Source Folder Endpoint",
        "description": "上传整個文件夹（逐文件）.",
        "operationId": "upload_data_source_folder_endpoint_api_v1_projects_data_sources_upload_folder_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_data_source_folder_endpoint_api_v1_projects_data_sources_upload_folder_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project Endpoint",
        "description": "获取单个项目.",
        "operationId": "get_project_endpoint_api_v1_projects__project_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project Endpoint",
        "description": "更新项目.",
        "operationId": "update_project_endpoint_api_v1_projects__project_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project Endpoint",
        "description": "删除项目.",
        "operationId": "delete_project_endpoint_api_v1_projects__project_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/archive": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Archive Project Endpoint",
        "description": "归档项目.",
        "operationId": "archive_project_endpoint_api_v1_projects__project_id__archive_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/restore": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Restore Project Endpoint",
        "description": "恢复项目.",
        "operationId": "restore_project_endpoint_api_v1_projects__project_id__restore_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_data_source_folder_endpoint_api_v1_projects_data_sources_upload_folder_post": {
        "properties": {
          "folder_name": {
            "type": "string",
            "title": "Folder Name"
          },
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          },
          "relative_paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Relative Paths"
          }
        },
        "type": "object",
        "required": [
          "folder_name",
          "files",
          "relative_paths"
        ],
        "title": "Body_upload_data_source_folder_endpoint_api_v1_projects_data_sources_upload_folder_post"
      },
      "DataSourceInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "total_files": {
            "type": "integer",
            "title": "Total Files"
          },
          "total_samples": {
            "type": "integer",
            "title": "Total Samples"
          }
        },
        "type": "object",
        "required": [
          "name",
          "total_files",
          "total_samples"
        ],
        "title": "DataSourceInfo",
        "description": "数据源目录."
      },
      "DataSourceUploadResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "total_files": {
            "type": "integer",
            "title": "Total Files"
          }
        },
        "type": "object",
        "required": [
          "name",
          "total_files"
        ],
        "title": "DataSourceUploadResponse",
        "description": "上传后返回的目录信息."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "priority": {
            "type": "string",
            "enum": [
              "normal",
              "high"
            ],
            "title": "Priority",
            "default": "normal"
          },
          "completion_rate": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Completion Rate",
            "default": 0
          },
          "available_samples": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Samples",
            "default": 0
          },
          "total_samples": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Samples",
            "default": 0
          },
          "data_source_folder": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Data Source Folder"
          }
        },
        "type": "object",
        "required": [
          "name",
          "data_source_folder"
        ],
        "title": "ProjectCreate",
        "description": "创建项目."
      },
      "ProjectListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "ProjectListResponse",
        "description": "分页响应."
      },
      "ProjectResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "priority": {
            "type": "string",
            "enum": [
              "normal",
              "high"
            ],
            "title": "Priority",
            "default": "normal"
          },
          "completion_rate": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Completion Rate",
            "default": 0
          },
          "available_samples": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Available Samples",
            "default": 0
          },
          "total_samples": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Samples",
            "default": 0
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "project_id",
          "created_at",
          "updated_at",
          "created_by",
          "updated_by",
          "is_archived"
        ],
        "title": "ProjectResponse",
        "description": "项目响应."
      },
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "normal",
                  "high"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "completion_rate": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Rate"
          },
          "available_samples": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Samples"
          },
          "total_samples": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Samples"
          }
        },
        "type": "object",
        "title": "ProjectUpdate",
        "description": "更新项目."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}